+++ classes/components/adapter/DbMysqlAdapter.class.php
@@ -5,7 +5,7 @@ use Flex\Annona\Db\DbMySqli;
 use Flex\Annona\Db\WhereHelper;
 use Flex\Annona\Db\WhereHelperInterface;
 
-class DbBaseAdapter extends BaseAdapter{
+class DbMysqlAdapter extends BaseAdapter{
     public const __version = '0.2';
     public WhereHelperInterface $whereHelper;
     public function __construct(